Add files via upload
You can directly open this file and execute the 3 lines from your command line to run the file.

# Note: 
As the code runs in real-time, the no of days parameter needs to be changed in order to get the same results which I have explained in the readme file. Other than that you can simply run these commands.